add sign in function add session (need test)
