+++ src/main/java/com/petplatform/controller/CommonController.java
@@ -5,6 +5,9 @@ import com.petplatform.service.CommonService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpSession;
+import java.security.NoSuchAlgorithmException;
+
 @RestController
 public class CommonController {
 
@@ -16,5 +19,11 @@ public class CommonController {
 
     @PostMapping("/user")
     UserDto postUserInfo(@RequestBody UserDto userDto){ return service.postUserInfo(userDto); }
+
+    @PostMapping("/sign_in")
+    UserDto signIn(@RequestBody UserDto userDto, HttpSession session) throws NoSuchAlgorithmException { return service.signIn(userDto, session); }
+
+    @GetMapping("/session")
+    UserDto getSession(HttpSession session){ return service.getSession(session); }
 
 }
