+++ src/main/java/com/petplatform/mapper/MemberMapper.java
@@ -1,18 +1,14 @@
 package com.petplatform.mapper;
 
-import com.petplatform.controller.MemberController;
 import com.petplatform.dto.MemberDto;
+import com.petplatform.dto.UserDto;
 import org.apache.ibatis.annotations.Mapper;
-import org.springframework.beans.factory.annotation.Autowired;
-
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
 
 @Mapper
 public interface MemberMapper {
 
     MemberDto selectMember(MemberDto info); // string id
 
-    void insertMember(MemberDto info);
+    void insertMember(UserDto info);
 }
